Add AsciiDoc specification document pipeline
Replaces the provisional Markdown spec (ife_spec.md) with a full AsciiDoc pipeline:

- `spec/ife_spec.adoc`: hand-written narrative with native `include::` directives
- `generator/emit/docs.py`: rewritten to emit one AsciiDoc file per item (layout/, constants/, provenance.adoc) instead of a single Markdown file
- `spec/build_document.sh`: one command builds HTML and PDF, fails on unresolved includes
- CI `document` job: renders both outputs and uploads as an artifact
- CI `tests` job: expanded to GCC, Clang, AppleClang, and MSVC matrices
- Orphan detection: `--check` and regeneration now fail/remove files the current spec no longer produces
- Conformance diagnostics cite stable clause ids (e.g. `ife-layer-extents`) instead of section numbers; `--validate` proves each id resolves in the narrative
- `<cmath>` explicitly included in the example for GCC 13 compatibility
- `IFE_EXPORT` removed from `AssociatedImages` typedef

+++ tests/ife_blocks_tests.cpp
@@ -709,14 +709,14 @@ void check_writes_exactly(const char* what, const CreateInfo& info) {
     }
     for (std::size_t i = 0; i < PAD; ++i) {
         if (buf[i] != POISON) {
-            std::fprintf(stderr, "FAIL: %s wrote %zu bytes BEFORE its offset\n", what, PAD - i);
+            std::fprintf(stderr, "FAIL: %s wrote %zu bytes BEFORE its offset\n", what, static_cast<std::size_t>(PAD - i));
             ++g_failures;
             break;
         }
     }
     for (std::size_t i = PAD + span; i < buf.size(); ++i) {
         if (buf[i] != POISON) {
-            std::fprintf(stderr, "FAIL: %s wrote past size_of() at +%zu\n", what, i - (PAD + span));
+            std::fprintf(stderr, "FAIL: %s wrote past size_of() at +%zu\n", what, static_cast<std::size_t>(i - (PAD + span)));
             ++g_failures;
             break;
         }
